feat - 로또 중복 숫자 유효성검사
